Clean up test
