+++ test/unitTests/logging/InformationMessageObserver.test.ts
@@ -5,8 +5,7 @@
 
 import { InformationMessageObserver } from '../../../src/observers/InformationMessageObserver';
 import { use as chaiUse, expect, should } from 'chai';
-import { vscode, Uri } from '../../../src/vscodeAdapter';
-import { getFakeVsCode, getNullWorkspaceConfiguration, getUnresolvedDependenices } from '../testAssets/Fakes';
+import { getUnresolvedDependenices, updateConfig, getVSCodeWithConfig } from '../testAssets/Fakes';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/observable/fromPromise';
 import 'rxjs/add/operator/timeout';
@@ -23,35 +22,11 @@ suite("InformationMessageObserver", () => {
     let commandDone = new Promise<void>(resolve => {
         signalCommandDone = () => { resolve(); };
     });
-    let vscode: vscode = getFakeVsCode();
+    let vscode = getVsCode();
     let infoMessage: string;
     let relativePath: string;
     let invokedCommand: string;
     let observer: InformationMessageObserver = new InformationMessageObserver(vscode);
-
-    vscode.window.showInformationMessage = async (message: string, ...items: string[]) => {
-        infoMessage = message;
-        return new Promise<string>(resolve => {
-            doClickCancel = () => {
-                resolve(undefined);
-            };
-
-            doClickOk = () => {
-                resolve(message);
-            };
-        });
-    };
-
-    vscode.commands.executeCommand = (command: string, ...rest: any[]) => {
-        invokedCommand = command;
-        signalCommandDone();
-        return undefined;
-    };
-
-    vscode.workspace.asRelativePath = (pathOrUri?: string, includeWorspaceFolder?: boolean) => {
-        relativePath = pathOrUri;
-        return relativePath;
-    };
 
     setup(() => {
         infoMessage = undefined;
@@ -66,34 +41,16 @@ suite("InformationMessageObserver", () => {
         let event = getUnresolvedDependenices("someFile");
 
         suite('Suppress Dotnet Restore Notification is true', () => {
-            setup(() => {
-                vscode.workspace.getConfiguration = (section?: string, resource?: Uri) => {
-                    return {
-                        ...getNullWorkspaceConfiguration(),
-                        get: <T>(section: string) => {
-                            return true;// suppress the restore information
-                        }
-                    };
-                };
-            });
+            setup(() => updateConfig(vscode, 'csharp', 'suppressDotnetRestoreNotification', true));
 
-            test('The information message is not shown', () => {    
+            test('The information message is not shown', () => {
                 observer.post(event);
                 expect(infoMessage).to.be.undefined;
             });
         });
-        
+
         suite('Suppress Dotnet Restore Notification is false', () => {
-            setup(() => {
-                vscode.workspace.getConfiguration = (section?: string, resource?: Uri) => {
-                    return {
-                        ...getNullWorkspaceConfiguration(),
-                        get: <T>(section: string) => {
-                            return false; // do not suppress the restore info
-                        }
-                    };
-                };
-            });
+            setup(() => updateConfig(vscode, 'csharp', 'suppressDotnetRestoreNotification', false));
 
             test('The information message is shown', async () => {
                 observer.post(event);
@@ -103,20 +60,49 @@ suite("InformationMessageObserver", () => {
                 await commandDone;
                 expect(invokedCommand).to.be.equal('dotnet.restore');
             });
-        
+
             test('Given an information message if the user clicks Restore, the command is executed', async () => {
                 observer.post(event);
                 doClickOk();
                 await commandDone;
                 expect(invokedCommand).to.be.equal('dotnet.restore');
             });
-        
+
             test('Given an information message if the user clicks cancel, the command is not executed', async () => {
                 observer.post(event);
                 doClickCancel();
                 await expect(Observable.fromPromise(commandDone).timeout(1).toPromise()).to.be.rejected;
                 expect(invokedCommand).to.be.undefined;
             });
-        });   
+        });
     });
+
+    function getVsCode() {
+        let vscode = getVSCodeWithConfig();
+        vscode.window.showInformationMessage = async (message: string, ...items: string[]) => {
+            infoMessage = message;
+            return new Promise<string>(resolve => {
+                doClickCancel = () => {
+                    resolve(undefined);
+                };
+
+                doClickOk = () => {
+                    resolve(message);
+                };
+            });
+        };
+
+        vscode.commands.executeCommand = (command: string, ...rest: any[]) => {
+            invokedCommand = command;
+            signalCommandDone();
+            return undefined;
+        };
+
+        vscode.workspace.asRelativePath = (pathOrUri?: string, includeWorspaceFolder?: boolean) => {
+            relativePath = pathOrUri;
+            return relativePath;
+        };
+
+        return vscode;
+    }
 });
